bot/requirements.txt
aiohttp==3.12.2
discord.py==2.5.2
